% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_CPT.R
\name{conduct_MCCPT}
\alias{conduct_MCCPT}
\title{Run a Monte-Carlo changepoints analysis on supplied paleoclimate record data}
\usage{
conduct_MCCPT(
  sites_data,
  age_lowerbound = 6000,
  age_upperbound = 22000,
  output_folder = paste0(getwd(), "/CPT_outputs/"),
  minseg_len = NULL,
  n_cpts = NULL,
  save = TRUE,
  rtn = TRUE,
  rev_y = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sites_data}{formatted paleoclimate data, output from import_sites.}

\item{age_lowerbound}{minimum age for the analysis in years BP.}

\item{age_upperbound}{maximum age for the analysis in years BP.}

\item{output_folder}{full file path to a folder where the outputs will be save.}

\item{minseg_len}{NULL or specify a minimum segment length. User will be prompted for each record if none is specified.}

\item{n_cpts}{number of changepoints to attempt. User will be prompted on a per-record basis if none is specified.}

\item{save}{TRUE/FALSE to save outputs to the output_folder.}

\item{rtn}{TRUE/FALSE to return changepoint data.}

\item{rev_y}{TRUE/FALSE to reverse y axis on plots}
}
\description{
Apply MCCPT to paleoclimate data
}
