% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-importers.R
\name{import_files}
\alias{import_files}
\title{Get all sites in the target folder}
\usage{
import_files(target_folder, verbose = TRUE)
}
\arguments{
\item{verbose}{TRUE/FALSE to print per-file update progress}

\item{folder_path}{a full path to the directory containing site xlsx files}
}
\description{
Scrape the supplied directory for sites for CPT analysis, and load them into a formatted list.
}
